src/main/java/edu/orange/internship/Downloader.java
package edu.orange.internship;

import java.io.BufferedInputStream;
import java.io.FileOutputStream;
import java.io.IOException;
import java.net.URL;
import java.util.regex.Matcher;
import java.util.regex.Pattern;

/**
 * Created by Mustafa on 8/1/2016.
 */
public class Downloader {

    private String urlString = "";
    public Boolean state = false;

    public Boolean download(String filename) throws IOException {
        if(state == false)
            return false;
        BufferedInputStream in = null;
        FileOutputStream fout = null;
        try {
            in = new BufferedInputStream(new URL(urlString).openStream());
            fout = new FileOutputStream(filename);

            final byte data[] = new byte[1024];
            int count;
            while ((count = in.read(data, 0, 1024)) != -1) {
                fout.write(data, 0, count);
            }
        }catch (IOException ioe){
            return false;
        }finally {
            if (in != null) {
                in.close();
            }
            if (fout != null) {
                fout.close();
            }
        }
        return true;
    }

    public Boolean enterUrl(String url) {
        urlString = url;
        String regex = "^(https?|ftp|file)://[-a-zA-Z0-9+&@#/%?=~_|!:,.;]*[-a-zA-Z0-9+&@#/%=~_|]";
        Pattern urlPattern = Pattern.compile(regex);
        Matcher urlMatcher = urlPattern.matcher(url);
        state = urlMatcher.matches();
        return state;
    }
}
